Create word count challenge
Intended as a more realistic example of a LINQ use case. Provides some good practice with grouping and filtering, along with some string methods.

The most confusing portion at first glance is likely line 20. This is to remove any punctuation from the string, which is done by filtering out punctuation characters and creating a new string. This could cause memory constraints with larger files, which may need to addressed in that case.